remove: section components
