--- components/SectionProject.vue
@@ -1,30 +1,0 @@
-<template>
-    <section class="mt-20 lg:mt-[188px]">
-        <div class="container relative">
-            <p class="text-white text-3xl md:text-6xl">
-                Il Padel World Club DAO <br>
-                mira a
-                <span class="inline-flex items-center gap-1 bg-secondary text-app-bg px-2 my-1">
-                    rivoluzionare
-                    <icon-rotate />
-                </span> <br>
-
-                l'esperienza del
-                <span class="inline-flex items-center gap-1 bg-secondary text-app-bg px-2 my-1">
-                    padel
-                    <icon-livello />
-                </span>
-                creando <br>
-                un circuito elitario di club <br>
-                diffusi <span class="inline-flex items-center gap-1 bg-secondary text-app-bg  px-2 my-1">
-                    globalmente
-                    <icon-globe />
-                </span> attraverso <br>
-                la tecnologia web3.
-            </p>
-            <span
-                class=" text-app-bg bg-[#D3E2E8] px-2 md:px-4 py-2 text-xl md:text-3xl absolute  -top-20 left-6 md:-left-2 rotate-[-5.45deg]">IL
-                PROGETTO</span>
-        </div>
-    </section>
-</template>
